Split thSetGetStyle to own Header
- prevent incomplete type thWindow error.
